ch4/am: refactor mpidig_am_comm_abort.c
