+++ src/mpid/ch4/src/ch4r_callbacks.h
@@ -20,9 +20,5 @@
 
 int MPIDIG_check_cmpl_order(MPIR_Request * req, MPIDIG_am_target_cmpl_cb target_cmpl_cb);
 void MPIDIG_progress_compl_list(void);
-int MPIDIG_comm_abort_origin_cb(MPIR_Request * sreq);
-int MPIDIG_comm_abort_target_msg_cb(int handler_id, void *am_hdr, void **data, size_t * p_data_sz,
-                                    int is_local, int *is_contig,
-                                    MPIDIG_am_target_cmpl_cb * target_cmpl_cb, MPIR_Request ** req);
 
 #endif /* CH4R_CALLBACKS_H_INCLUDED */
